FIX calculation of block match offsets
Introduce block_xy_similarity in order to more easily report positions
offsets corresponding to correlation values. In other words, we feed
in a large image and a small block from a similar image and a position
of a roughly corresponding block in the reference. We get back a
correlation matrix whose indices correspond to offsets relative to the
entire reference, but we also get back an xoff and yoff array that
translate these offsets with respect to the block in the reference.