FIX consistent ordering of xoff and yoff
Was getting the x and y offsets corresponding to the correlation
matrix out of order because of inconsistency between function calls
and outputs.

Also, now look in larger range of z values.

And save more information about the correlations matrix into the mat
file--particularly, where the blocks were and the best x and y offsets