ADD option to use previous search range for xy
The option is called useXYSearchRangeFromDate and if you give it a
date, it will use that date to constrain the possible xy positions of
the registration. The new options searchRangeXMargin and
searchRangeYMargin determine how far around the search range center
you should look for each block.

Could consider also passing the correlation values in to use to
determine outliers